Hit a snag, need to look up some things
I want to put a complicated iterative for each loop at the presently commented section of code.  But I need to look up how to do that in Python so that it changes the variable on each loop.  This was the whole reason for wanting to do this script, so, stay tuned! <3
- Scaerie