Made some progress :)
Sorry, I know it's going slow.  I've been learning a *lot* lately :)  I've put some more work into the experimental branch here.  Just need to put a for loop in to cycle through the pet files, and then put a GUI on top so it's not a pain to use ♥